+++ src/string_search/find_char_not.c
@@ -6,7 +6,7 @@
 /*   By: rjw <rjw@student.codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/09/11 20:30:39 by rjw           #+#    #+#                 */
-/*   Updated: 2025/03/05 21:37:59 by rjw           ########   odam.nl         */
+/*   Updated: 2025/04/11 02:04:32 by rjw           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ size_t	find_set_not(t_cchr *str, char *set)
 		c = *tmp_str;
 	}
 	if (ft_strchr(set, c) == NULL)
-		return (tmp_str - str);
+		return ((size_t)(tmp_str - str));
 	return (0);
 }
 
@@ -47,7 +47,7 @@ size_t	find_char_not(t_cchr *str, t_cchr c1)
 	}
 	if (*tmp_str != c1)
 	{
-		return (tmp_str - str);
+		return ((size_t)(tmp_str - str));
 	}
 	return (0);
 }
@@ -69,7 +69,7 @@ size_t	find_char2_not(t_cchr *str, t_cchr c1, t_cchr c2)
 		c = *tmp_str;
 	}
 	if (c != c1 && c != c2)
-		return (tmp_str - str);
+		return ((size_t)(tmp_str - str));
 	return (0);
 }
 
@@ -90,6 +90,6 @@ size_t	find_char3_not(t_cchr *str, t_cchr c1, t_cchr c2, t_cchr c3)
 		c = *tmp_str;
 	}
 	if (c != c1 && c != c2 && c != c3)
-		return (tmp_str - str);
+		return ((size_t)(tmp_str - str));
 	return (0);
 }
